Backend/requirements.txt
# Flask Backend Dependencies
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# PyTorch and Transformers for model inference
torch>=2.0.0
transformers>=4.35.0
safetensors>=0.4.0

# Hugging Face Hub for model downloading
huggingface-hub>=0.19.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
